SECOAUTH-285: Created OAuth2ExpressionParser that automatically wraps a Spring Expression with a method that provides meaningful errors.
This allows Spring Security OAuth to check for errors and throw Excepions when an error occurs while ensuring that the exception is only
thrown when necessary. For example, if an expression has an or statment we do not want to throw the Exception if only part of the expression is false.